Downloads/BlueTeamBotLite/BlueTeamBotLite/bluebot-lite/requirements.txt
python-nmap==0.7.1
PyYAML==6.0.2
pandas==2.2.2
reportlab==4.2.2
tabulate==0.9.0
